requirements.txt
microsoft-autogen==0.2.18
openai>=1.0.0
python-dotenv==1.0.0
pydantic==2.5.0
asyncio
json-schema
typing-extensions